+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Restructure project and harmonize build
 - Include client IP in JIT rules
 - Enable individual Gen1/Gen2 VMs (#1528)
+- Added some help to discover the error cause when `New-LabAzureResourceGroupDeployment` failed 
 
 ### Bugs